+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/prepublish/PrePublishTimerBean.java
@@ -45,8 +45,18 @@ public class PrePublishTimerBean {
 
     @Timeout
     public void performPrePublish() {
+
+        final int maxEventsPublishedPerIteration = prePublishTimerConfig.getMaxEventsPublishedPerIteration();
+
+        int numberOfEventsPublished = 1;
         while (prePublishProcessor.prePublishNextEvent()) {
+
             timerServiceManager.cancelOverlappingTimers(TIMER_JOB_NAME, THRESHOLD, timerService);
+            numberOfEventsPublished++;
+
+            if (numberOfEventsPublished > maxEventsPublishedPerIteration) {
+                break;
+            }
         }
     }
 }